+++ src/components/pages/Login/Login.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { Login } from '../../containers/Login/Login';
+import { Login } from './Login';
 
 describe('Tooltip test', () => {
   const createLogin = () => <Login />;
